feat(native): commit remaining verified native delta — HiZ mount refinements, chart text density, telemetry, dashboard pointer refinements
Working-tree snapshot that compiles green (lib 430, bin 165, all suites):
HiZ production mount refinements + chart text density fixture test +
material fastpath polish + dashboard pointer/table-action refinements +
dynamic playback and state-ops tweaks + cascade/mesh shader updates.
Mixed provenance from tonight's lanes and the parallel session; all
tests green in this exact state, so the snapshot is the coherent
verified point.

+++ packages/deep-engine-native/tests/support/lod_draw_renderer.rs
@@ -128,7 +128,7 @@ pub async fn render(
         &culling,
         lod.as_ref(),
         &pipelines,
-        u8::MAX,
+        u16::MAX,
     );
     encode_mesh_passes(
         &mut encoder,
